Restructure example files
