+++ examples/javascript/as_promises.js
