+++ examples/javascript/charge_schedule.js
