+++ examples/javascript/create_customer.js
