+++ examples/javascript/create_link.js
