+++ examples/javascript/create_source_charge.js
