+++ examples/javascript/create_transfer.js
