+++ examples/javascript/destroy_customer.js
