+++ examples/javascript/list_charges.js
