+++ examples/javascript/list_schedules.js
