+++ examples/javascript/retrieve_balance.js
