+++ examples/javascript/retrieve_customer.js
