+++ examples/javascript/update_customer.js
