+++ examples/typescript/index.ts
@@ -2,7 +2,7 @@
 import * as dotenv from 'dotenv';
 dotenv.config();
 
-import omiseNode from '../../index'
+import omiseNode from '../../index';
 
 const omise = omiseNode({
   publicKey: process.env.OMISE_PUBLIC_KEY,
@@ -10,4 +10,4 @@ const omise = omiseNode({
   host: process.env.OMISE_NODE_HOST,
 });
 
-export default omise
+export default omise;
