+++ examples/typescript/list_customers.ts
@@ -1,5 +1,5 @@
 import omise from "./index";
-import { Customers } from "../index";
+import { Customers } from "../../types";
 
 omise.customers.list({}, (err: Error | null, resp: Customers.ICustomerList) => {
   if (err) {
