+++ examples/typescript/list_recipient.ts
@@ -1,5 +1,5 @@
 import omise from "./index";
-import { Recipients } from "../index";
+import { Recipients } from "../../types";
 
 omise.recipients.list({limit: 5}, (err: Error | null, resp: Recipients.IRecipientList) => {
   if (err) {
