+++ examples/typescript/retrieve_balance.ts
@@ -1,5 +1,5 @@
 import omise from './index';
-import { Balance } from '../index';
+import { Balance } from '../../types';
 
 omise.balance.retrieve((err: Error | null, resp: Balance.IBalance) => {
   if (err) {
